v2022

+++ inventorview/tools/chocolateyinstall.ps1
@@ -1,17 +1,30 @@
 ﻿$ErrorActionPreference = 'Stop';
 
-$url             = "Z:\inventorview\inventor_view_2022_english_win_64bit_dlm.sfx.exe"
-#$url             = 'https://download.autodesk.com/us/support/files/autodesk_inventor_view_2022/inventor_view_2022_english_win_64bit_dlm.sfx.exe'
+$url             = 'https://download.autodesk.com/us/support/files/autodesk_inventor_view_2022/inventor_view_2022_english_win_64bit_dlm.sfx.exe'
 $checksum        = '94C4E11509DEA6F7857B936920E627D27160EE2EA148B72CFCF1B48477B6C94B'
 $unzip           = Join-Path $env:TEMP 'Inventor_View_2022_English_Win_64bit_DLM'
-$regkey          = 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\{7F4DD591-2664-0004-0000-7107D70F3DB4}'
-$regkeylp        = 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\{7F4DD591-2664-0004-1033-7107D70F3DB4}'
 
+$regkey2021lp    = 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\{7F4DD591-2564-0004-1033-7107D70F3DB4}'
+$regkey2022      = 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\{7F4DD591-2664-0004-0000-7107D70F3DB4}'
+$regkey2022lp    = 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\{7F4DD591-2664-0004-1033-7107D70F3DB4}'
+
+#Uninstall old versions first
+$packagename     = 'Inventor View 2021'
+$productcode     = '{7F4DD591-2564-0004-0000-7107D70F3DB4}'
+$regkey          = "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\$productcode"
+if (Test-path $regkey) { Uninstall-ChocolateyPackage -PackageName "$packagename" -FileType "msi" -SilentArgs "$productcode /qn /norestart" }
+
+$packagename     = 'Language Pack 2021'
+$productcode     = '{7F4DD591-2564-0004-0000-7107D70F3DB4}'
+$regkey          = "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\$productcode"
+if (Test-path $regkey) { Uninstall-ChocolateyPackage -PackageName "$packagename" -FileType "msi" -SilentArgs "$productcode /qn /norestart" }
+
+#Extract first
 $packageArgsUnzip = @{
-  packageName    = 'IVV Unzip'
+  packageName    = 'Unzip'
   fileType       = 'exe'
   url            = $url
-  softwareName   = 'IVV Unzip*'
+  softwareName   = 'Unzip*'
   checksum       = $checksum
   checksumType   = 'sha256'
   silentArgs     = "-suppresslaunch -d $env:TEMP"
@@ -19,13 +32,11 @@ $packageArgsUnzip = @{
 }
 Install-ChocolateyPackage @packageArgsUnzip
 
-#remove any reboot requests that may block the installation
+#Remove any reboot requests that may block the installation
 $RegRebootRequired = "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\WindowsUpdate\Auto Update\RebootRequired"
-if (Test-path $RegRebootRequired)
-{
-    Remove-Item -Path $RegRebootRequired
-}
+if (Test-path $RegRebootRequired) { Remove-Item -Path $RegRebootRequired }
 
+#Setup.exe not silent, installing all modules one by one
 $adsso           = Join-Path $unzip 'x64\AdSSO\AdSSO.msi'
 $packageArgsAdSSO  = @{
   packageName    = 'Autodesk Single Sign On Component'
@@ -47,6 +58,7 @@ $packageArgsIVV  = @{
   validExitCodes = @(0, 3010, 1641)
 }
 Install-ChocolateyInstallPackage @packageArgsIVV
+if (Test-path $regkey2022) { Set-Itemproperty -path $regkey2022 -Name 'SystemComponent' -value '0' -Type dword }
 
 $ivvlp           = Join-Path $unzip 'x64\en-US\IVV\inventorViewLP.msi'
 $packageArgsIVVLP  = @{
@@ -58,6 +70,7 @@ $packageArgsIVVLP  = @{
   validExitCodes = @(0, 3010, 1641)
 }
 Install-ChocolateyInstallPackage @packageArgsIVVLP
+if (Test-path $regkey2022lp) { Set-Itemproperty -path $regkey2022lp -Name 'SystemComponent' -value '0' -Type dword }
 
 $adapp           = Join-Path $unzip 'x86\ADSKAPP\AdApplicationManager-installer.exe'
 $packageArgsAdApp = @{
@@ -69,10 +82,3 @@ $packageArgsAdApp = @{
   validExitCodes = @(0, 3010, 1641)
 }
 Install-ChocolateyInstallPackage @packageArgsAdApp
-
-#msi installers are missing uninstall entry, creating it manually
-if (Test-path $regkey)
-{
-Set-Itemproperty -path $regkey -Name 'SystemComponent' -value '0' -Type dword
-Set-Itemproperty -path $regkeylp -Name 'SystemComponent' -value '0' -Type dword
-}
